refactor(images): refactor images catalog in source
